rustc-coverage-tests/requirements.txt
pyyaml
tabulate
